Another take on supporting trailing response headers on the client
Added a generic parameter to pass along the knowledge of what kinds of objects
are known to be flowing through the streams. The wrapData/wrapTrailingHeaders
methods still feel hacky, but at least now they're pure functions and applied
immediately when the respective frames come in, so I think this is at least
better than the promise/future approach.

There is 1 failing test that appears to be a bug: it seems IncomingStreamBuffer
is not pulled when it has emitted all data but the trailing header is still
pending.